+++ system/info_cpu_linux.go
@@ -10,6 +10,7 @@ package system
 import (
 	"bufio"
 	"errors"
+	"io/ioutil"
 	"strconv"
 	"strings"
 	"time"
@@ -24,6 +25,14 @@ var procStatFile = "/proc/stat"
 
 // Path to file with info about CPU
 var cpuInfoFile = "/proc/cpuinfo"
+
+// Files with CPU info
+var (
+	cpuPossibleFile = "/sys/devices/system/cpu/possible"
+	cpuPresentFile  = "/sys/devices/system/cpu/present"
+	cpuOnlineFile   = "/sys/devices/system/cpu/online"
+	cpuOfflineFile  = "/sys/devices/system/cpu/offline"
+)
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -101,6 +110,40 @@ func GetCPUInfo() ([]*CPUInfo, error) {
 	defer closer()
 
 	return parseCPUInfo(s)
+}
+
+// GetCPUCount returns info about CPU
+func GetCPUCount() (CPUCount, error) {
+	possible, err := ioutil.ReadFile(cpuPossibleFile)
+
+	if err != nil {
+		return CPUCount{}, err
+	}
+
+	present, err := ioutil.ReadFile(cpuPresentFile)
+
+	if err != nil {
+		return CPUCount{}, err
+	}
+
+	online, err := ioutil.ReadFile(cpuOnlineFile)
+
+	if err != nil {
+		return CPUCount{}, err
+	}
+
+	offline, err := ioutil.ReadFile(cpuOfflineFile)
+
+	if err != nil {
+		return CPUCount{}, err
+	}
+
+	return CPUCount{
+		Possible: parseCPUCountInfo(string(possible)),
+		Present:  parseCPUCountInfo(string(present)),
+		Online:   parseCPUCountInfo(string(online)),
+		Offline:  parseCPUCountInfo(string(offline)),
+	}, nil
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -246,3 +289,14 @@ func parseCPUInfo(s *bufio.Scanner) ([]*CPUInfo, error) {
 }
 
 // codebeat:enable[LOC,ABC,CYCLO]
+
+// parseCPUCountInfo parses CPU count data
+func parseCPUCountInfo(data string) uint32 {
+	startNum := strings.Trim(strutil.ReadField(data, 0, false, "-"), "\n\r")
+	endNum := strings.Trim(strutil.ReadField(data, 1, false, "-"), "\n\r")
+
+	start, _ := strconv.ParseUint(startNum, 10, 32)
+	end, _ := strconv.ParseUint(endNum, 10, 32)
+
+	return uint32(end-start) + 1
+}